move napi_build_version to separate test fixture refactored rebuild helper
