+++ test/fixture/napi-build-version/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "workspace-app",
+  "productName": "Workspace App",
+  "version": "1.0.0",
+  "description": "",
+  "main": "src/index.js",
+  "keywords": [],
+  "author": "",
+  "license": "MIT",
+  "dependencies": {
+    "sqlite3": "5.1.6"
+  }
+}
